added misc stuff

+++ src/main/java/com/gamesense/client/command/commands/HClipCommand.java
@@ -22,7 +22,7 @@ public class HClipCommand extends Command {
             String main = message[0];
 
             try {
-                amount = Integer.parseInt(main);
+                amount = Double.parseDouble(main);
                 MessageBus.sendCommandMessage(ModuleManager.getModule(ColorMain.class).getEnabledColor() + "Clipped the player " + main + " blocks forward.", true);
 
                 final Vec3d dir = new Vec3d(Math.cos((mc.player.rotationYaw + 90f) * Math.PI / 180.0f), 0, Math.sin((mc.player.rotationYaw + 90f) * Math.PI / 180.0f));
